src/driver/pyproject.toml

[build-system]
requires = ["uv_build>=0.8.20,<0.9.0"]
build-backend = "uv_build"

[project]
name = "alpasim_driver"
version = "0.2.0"
description = "Adapters for open-source drivers in Alpasim"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "vam",
  "einops",
  "numpy",
  "torch",
  "pillow",
  "rich",
  "click",
  "grpcio",
  "hydra-core",
  "omegaconf",
  "alpasim-grpc"
]

  [project.scripts]
  vam_driver = "alpasim_driver.main:main"

[tool.uv.sources]
vam = { git = "https://github.com/valeoai/VideoActionModel.git" }
alpasim-grpc = { workspace = true }
